refactor: Order 리팩토링
