--- src/main/java/kitchenpos/dao/OrderDao.java
@@ -1,19 +1,0 @@
-package kitchenpos.dao;
-
-import java.util.List;
-import java.util.Optional;
-
-import kitchenpos.domain.Order;
-
-@Deprecated
-public interface OrderDao {
-    Order save(Order entity);
-
-    Optional<Order> findById(Long id);
-
-    List<Order> findAll();
-
-    boolean existsByOrderTableIdAndOrderStatusIn(Long orderTableId, List<String> orderStatuses);
-
-    boolean existsByOrderTableIdInAndOrderStatusIn(List<Long> orderTableIds, List<String> orderStatuses);
-}
