--- src/main/java/kitchenpos/dao/TableGroupDao.java
@@ -1,15 +1,0 @@
-package kitchenpos.dao;
-
-import java.util.List;
-import java.util.Optional;
-
-import kitchenpos.domain.TableGroup;
-
-@Deprecated
-public interface TableGroupDao {
-    TableGroup save(TableGroup entity);
-
-    Optional<TableGroup> findById(Long id);
-
-    List<TableGroup> findAll();
-}
