+++ src/main/java/kitchenpos/domain/OrderLineItem.java
@@ -1,40 +1,70 @@
 package kitchenpos.domain;
 
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+@Entity
+@Table(name = "order_line_item")
 public class OrderLineItem {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long seq;
-    private Long orderId;
-    private Long menuId;
-    private long quantity;
 
-    public Long getSeq() {
-        return seq;
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "order_id")
+    private Order order;
+
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "menu_id")
+    private Menu menu;
+
+    private Integer quantity;
+
+    protected OrderLineItem() {
     }
 
-    public void setSeq(final Long seq) {
-        this.seq = seq;
-    }
-
-    public Long getOrderId() {
-        return orderId;
-    }
-
-    public void setOrderId(final Long orderId) {
-        this.orderId = orderId;
-    }
-
-    public Long getMenuId() {
-        return menuId;
-    }
-
-    public void setMenuId(final Long menuId) {
-        this.menuId = menuId;
-    }
-
-    public long getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(final long quantity) {
+    public OrderLineItem(Order order, Menu menu, Integer quantity) {
+        setOrder(order);
+        this.menu = menu;
         this.quantity = quantity;
+    }
+
+    public static List<OrderLineItem> of(Order order, Map<Menu, Integer> menus) {
+        return menus.entrySet()
+            .stream()
+            .map(entry -> new OrderLineItem(order, entry.getKey(), entry.getValue()))
+            .collect(Collectors.toList());
+    }
+
+    public void setOrder(Order newOrder) {
+        if (Objects.nonNull(order)) {
+            order.getOrderLineItems().remove(this);
+        }
+
+        order = newOrder;
+
+        if (!order.getOrderLineItems().contains(this)) {
+            order.getOrderLineItems().add(this);
+        }
+    }
+
+    public String getMenuName() {
+        return menu.getName();
+    }
+
+    public Integer getQuantity() {
+        return quantity;
     }
 }
