+++ src/main/java/kitchenpos/domain/OrderRepository.java
@@ -2,5 +2,5 @@ package kitchenpos.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Order2, Long> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
