+++ src/main/java/kitchenpos/ui/dto/OrderRequest.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 import kitchenpos.domain.Menu;
-import kitchenpos.domain.Order2;
+import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderTable2;
 import kitchenpos.exception.EntityNotFoundException;
 
@@ -25,8 +25,8 @@ public class OrderRequest {
 			.collect(Collectors.toList());
 	}
 
-	public Order2 toOrder(OrderTable2 orderTable, List<Menu> menus) {
-		return new Order2(orderTable, getMenusWithQuantity(menus));
+	public Order toOrder(OrderTable2 orderTable, List<Menu> menus) {
+		return new Order(orderTable, getMenusWithQuantity(menus));
 	}
 
 	private Map<Menu, Integer> getMenusWithQuantity(List<Menu> menus) {
